project/src/requirements.txt

lightrag-hku
aiofiles
asyncpg           
bcrypt            
cryptography      
exceptiongroup    
faiss-cpu         
fastapi           
importlib_metadata
ipykernel         
langchain-community
neo4j             
ollama            
openai            
passlib           
pickleshare       
pip               
PyJWT             
PyMuPDF           
python-jose       
python-multipart  
uvicorn           
wheel             